test: give the GUI-less configuration a unit suite of its own
A CONTOUR_FRONTEND_GUI=OFF build had no unit tests at all. Not by decision:
every test lived in contour_gui_test, whose runner constructs a
QGuiApplication, and that build has no Qt. So the layers that configuration
consists of -- the configuration model, the command vocabulary and the CLI --
were built there and asserted nowhere.

contour_test is those twelve suites with a runner that constructs nothing:
2189 assertions across 79 cases, linking contour_command and contour_cli and
no Qt. The headless CI job now builds and runs it instead of only checking
that the binary links.

Two fixtures had to stop depending on Qt to get there: crispy has no temp-dir
helper and QTemporaryDir was doing that job, so test/TempDir.h does it with
std::filesystem, and the Qt-free half of GuiTestFixtures.h moves to
test/CoreFixtures.h.

Config_test stays with the GUI tests on purpose: besides the configuration
model it also asserts the Qt colour-scheme mapping and uses qputenv, so it is
not Qt-free however Qt-free the model it tests is.
